backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.16.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
gunicorn>=21.0.0
dill>=0.3.7
xgboost>=2.0.0
requests>=2.31.0